+++ src/main/java/io/nonamuckja/backend/security/service/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package io.nonamuckja.backend.service;
+package io.nonamuckja.backend.security.service;
 
 import java.util.Optional;
 import java.util.stream.Collectors;
