+++ src/main/java/io/nonamuckja/backend/security/service/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package io.nonamuckja.backend.security;
+package io.nonamuckja.backend.security.service;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
